+++ lumin/nn/models/blocks/body.py
@@ -1,7 +1,6 @@
 from typing import Optional, Callable, Any, List, Dict
 import numpy as np
 from functools import partial
-from abc import abstractmethod
 
 import torch.nn as nn
 import torch
@@ -20,20 +19,6 @@ class AbsBody(AbsBlock):
                  lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False):
         super().__init__(lookup_init=lookup_init, freeze=freeze)
         self.n_in,self.feat_map,self.lookup_act = n_in,feat_map,lookup_act
-
-    @abstractmethod
-    def forward(self, x:Tensor) -> Tensor:
-        r'''
-        Pass tensor through body
-
-        Arguments:
-            x: incoming tensor
-        
-        Returns
-            Resulting tensor
-        '''
-
-        pass
     
 
 class FullyConnected(AbsBody):